+++ Forge/changelog.md
@@ -1,2 +1,4 @@
 1.2.0:
 * Added Cotton, Scythes and Seed Satchel
+* Added Scythe recipes
+* Added "Grim Reaper" advancement